third_party/blink/web_tests/http/tests/devtools/elements/styles-2/get-set-stylesheet-text-expected.txt

Tests that WebInspector.CSSStyleSheet methods work as expected.


Running: testSetText
=== Original stylesheet text: ===
body {
    font-size: 12px;
    foo: bar !

    important /* !important property not recognized by WebCore */;
    height: 100% !important;
    colorz: "!important";
    important:!important;
    color: "badcolor" ! important /* good property with strange value */;
}

@media {
    /* @media rule with an empty media list */
}

/* comment before selector */body.main1/* comment after selector */{/* comment */color: #F00BAA;zoo:moo /* not an !important unrecognized property */}/* comment */

body.main2{background: green /* value !important comment */ !important /* no semicolon, very !important */}


Running: testNewElementStyles
=== Matched rules for h1 after setText() ===

rule
h1: [user-agent]
['display':'block'] @[undefined-undefined] 
['font-size':'2em'] @[undefined-undefined] 
['margin-block-start':'0.67em'] @[undefined-undefined] 
['margin-block-end':'0.67em'] @[undefined-undefined] 
['margin-inline-start':'0px'] @[undefined-undefined] 
['margin-inline-end':'0px'] @[undefined-undefined] 
['font-weight':'bold'] @[undefined-undefined] 

rule
address, blockquote, center, div, figure, figcaption, footer, form, header, hr, legend, listing, main, p, plaintext, pre, summary, xmp, article, aside, h1, h2, h3, h4, h5, h6, hgroup, nav, section, search, table, caption, colgroup, col, thead, tbody, tfoot, tr, td, th, dir, dd, dl, dt, menu, ol, ul, li, bdi, output, [dir="ltr" i], [dir="rtl" i], [dir="auto" i]: [user-agent]
['unicode-bidi':'isolate'] @[undefined-undefined] 

rule
h1: [regular]
['COLOR':'Red'] @[0:5-0:16] 
['color':'red'] @[undefined-undefined] 
